+++ version/internal/types.go
@@ -1,16 +1,10 @@
 package internal
 
-// Version defines the response version.
-type Version struct {
-	Build     int
-	Mcversion string
-	Modified  float64
-	Version   string
-}
+// Root defines the root response element for forge versions.
+type Root map[string][]string
 
-// Root defines the root response element.
-type Root struct {
-	Homepage string
-	Number   map[string]Version
-	Promos   map[string]int
+// Slim defines the root response element for the promotions.
+type Slim struct {
+	Homepage string            `json:"homepage"`
+	Promos   map[string]string `json:"promos"`
 }